Added a process check module
VynaeIntegrityCheck creates a scheduled task that compares current processes to a control list created when it is first run. It then notifies the user of new or missing processes using compare-object and out-gridview. More modules will be added over time, and most future expansions to the Vynae project will be in this form.